[APM] Add NP feature catalog registration (#51765)
* Move register_feature.js to legacy_register_feature.js
* Add feature catalog registration to NP plugin
* Add plugin dependencies in kibana.json

Fixes #51542.
