+++ x-pack/plugins/apm/kibana.json
@@ -3,12 +3,7 @@
   "server": true,
   "version": "8.0.0",
   "kibanaVersion": "kibana",
-  "configPath": [
-    "xpack",
-    "apm"
-  ],
+  "configPath": ["xpack", "apm"],
   "ui": false,
-  "requiredPlugins": [
-    "apm_oss"
-  ]
+  "requiredPlugins": ["apm_oss", "data", "home"]
 }
